Skille skjema for kontoinfo ut i egen rneder
